PSP-2306 Lease Deposits (#511)
* Backend API mappings for Lease Security Deposits

* Add generic helpers to render Table cell values

* Add common section header UI

* Create skeleton form for Lease Deposits

* Include security deposit information on Lease API

* Implement derived field in security deposits table

* Add deposit return notes section

* Unit tests for lease deposits

* PR feedback

+++ backend/entities/SecurityDeposit.cs
@@ -77,7 +77,7 @@ namespace Pims.Dal.Entities
         /// get/set - The annual interest rate.
         /// </summary>
         [Column("ANNUAL_INTEREST_RATE")]
-        public int AnnualInterestRate { get; set; }
+        public decimal AnnualInterestRate { get; set; }
 
         #endregion
